requirements.txt
pandas
openai
langchain
langchain-openai
langchain-community
pypdf
faiss-cpu
python-dotenv
jinja2
jupyter
pyyaml
tiktoken
openpyxl
langchain-google-genai
google-generativeai
